Added security authentication for the user
The security pass now checks whether the person who logs in has an
account. Spring connects to the DB and checks if the credentials and
name match. The logged in user gets his corresponding role and is logged
into the app.

WIP

TODO: check how to get the whole user object from the DB, for the only
the theoretical part is written down, but no the full code.